refactor(BE): User 서비스 디렉토리 생성
- auth_service.py 파일과 로직 분리
