+++ app/backend/src/main/core/auth/dependencies.py
@@ -2,10 +2,9 @@ from fastapi import Depends, HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 from fastapi.security import OAuth2PasswordBearer
 
-from src.main.domains.user.schemas.user.user_create import UserCreate
-
 from .jwt import verify_token
-from src.main.domains.user.services import UserService
+from src.main.domains.user.schemas.user.user_create import UserCreate
+from src.main.domains.user.service.user_service import UserService
 from src.main.db.deps import get_db
